refactor: rename providers dir to contexts dir
